Fix overflow in type cast for GUPS benchmark (#58)
* Fix overflow in type cast for GUPS benchmark

Changed size_t n = (size_t)(1 << logn) to size_t n = ((size_t)1) << logn
to prevent integer overflow when logn >= 31. The original code would 
perform the left shift on an int (1) before casting to size_t, causing
overflow. Now the cast happens first, allowing proper handling of large
values like n = 2^31.

Fixes #57

+++ posts/gups/gups.cu
@@ -390,7 +390,7 @@ int main(int argc, char* argv[])
     }
   }
 
-  size_t n_shmem = (size_t)1 << logn_shmem;
+  size_t n_shmem = 1UL << logn_shmem;
 
   int ndev;
   CUDA_RT_CALL(cudaGetDeviceCount(&ndev));
@@ -458,8 +458,8 @@ int main(int argc, char* argv[])
         prop.sharedMemPerBlockOptin));
   }
 
-  size_t n = (size_t)(1 << logn);
-  size_t working_set = (size_t)n * 100 / occupancy;
+  size_t n = 1UL << logn;
+  size_t working_set = n * 100 / occupancy;
   if (accesses_per_elem == -1) {
     if (test_type == UPDATE || test_type == UPDATE_NO_LOOP
         || test_type == READ_WRITE) {
@@ -596,3 +596,4 @@ int main(int argc, char* argv[])
     CUDA_RT_CALL(cudaFree(d_t));
   return 0;
 }
+
